Der Extractor und der HttpPull arbeiten jetzt zusammen. Daten können jetzt über die DataPull Klasse als Extractor bezogen werden.
