+++ app/src/main/java/de/codecrops/vertretungsplangymwen/MainActivity.kt
@@ -2,10 +2,9 @@ package de.codecrops.vertretungsplangymwen
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-import de.codecrops.vertretungsplangymwen.network.HttpPull
-import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
